add e2e ci

+++ tools/java_driver/src/test/java/com/alibaba/neug/driver/JavaDriverE2ETest.java
@@ -47,10 +47,10 @@ public class JavaDriverE2ETest {
             try (Session session = driver.session();
                     ResultSet resultSet = session.run("RETURN 1 AS value")) {
                 assertTrue(resultSet.next());
-                assertEquals(1, resultSet.getInt("value"));
-                assertEquals(1, resultSet.getObject(0));
+                assertEquals(1L, resultSet.getLong("value"));
+                assertEquals(1L, resultSet.getObject(0));
                 assertFalse(resultSet.wasNull());
-                assertEquals(Types.INT32, resultSet.getMetaData().getColumnType(0));
+                assertEquals(Types.INT64, resultSet.getMetaData().getColumnType(0));
                 assertEquals("value", resultSet.getMetaData().getColumnName(0));
                 assertFalse(resultSet.next());
             }
